moved `PydanticIndicators` and `PydanticSubfield` into components.py
